Abstract out agent join
